Embed CreatePasswordTableViewController in CreatePasswordViewController
