+++ Kickstarter-iOS/Views/Controllers/CreatePasswordTableControllerTests.swift
@@ -3,7 +3,7 @@ import Prelude
 @testable import Library
 import XCTest
 
-final class CreatePasswordViewControllerTests: TestCase {
+final class CreatePasswordTableViewControllerTests: TestCase {
   override func setUp() {
     super.setUp()
 
@@ -21,7 +21,7 @@ final class CreatePasswordViewControllerTests: TestCase {
   func testCreatePassword() {
     combos(Language.allLanguages, Device.allCases).forEach { language, device in
       withEnvironment(language: language) {
-        let controller = CreatePasswordViewController.instantiate()
+        let controller = CreatePasswordTableViewController.instantiate()
         let (parent, _) = traitControllers(device: device, orientation: .portrait, child: controller)
 
         FBSnapshotVerifyView(parent.view, identifier: "lang_\(language)_device_\(device)")
@@ -30,27 +30,28 @@ final class CreatePasswordViewControllerTests: TestCase {
   }
 
   func testNewPasswordTextFieldTargetActions() {
-    let vc = MockCreatePasswordViewController()
+    let vc = MockCreatePasswordTableViewController()
     let actions = textFieldTargetActions(for: vc, row: .newPassword)
 
-    XCTAssertEqual(vc, actions[0].target as? MockCreatePasswordViewController)
+    XCTAssertEqual(vc, actions[0].target as? MockCreatePasswordTableViewController)
     XCTAssertTrue(sel_isEqual(actions[0].action, #selector(vc.newPasswordTextFieldChanged(_:))))
-    XCTAssertEqual(vc, actions[1].target as? MockCreatePasswordViewController)
+    XCTAssertEqual(vc, actions[1].target as? MockCreatePasswordTableViewController)
     XCTAssertTrue(sel_isEqual(actions[1].action, #selector(vc.newPasswordTextFieldDidReturn(_:))))
   }
 
   func testNewPasswordConfirmationTextFieldTargetActions() {
-    let vc = MockCreatePasswordViewController()
+    let vc = MockCreatePasswordTableViewController()
     let actions = textFieldTargetActions(for: vc, row: .confirmNewPassword)
 
-    XCTAssertEqual(vc, actions[0].target as? MockCreatePasswordViewController)
+    XCTAssertEqual(vc, actions[0].target as? MockCreatePasswordTableViewController)
     XCTAssertTrue(sel_isEqual(actions[0].action, #selector(vc.newPasswordConfirmationTextFieldChanged(_:))))
-    XCTAssertEqual(vc, actions[1].target as? MockCreatePasswordViewController)
+    XCTAssertEqual(vc, actions[1].target as? MockCreatePasswordTableViewController)
     XCTAssertTrue(sel_isEqual(actions[1].action, #selector(vc.newPasswordConfirmationTextFieldDidReturn(_:))))
   }
 }
 
-fileprivate final class MockCreatePasswordViewController: UIViewController, CreatePasswordViewControllerType {
+fileprivate final class MockCreatePasswordTableViewController: UITableViewController,
+CreatePasswordTableViewControllerType {
   @objc func newPasswordTextFieldChanged(_ sender: UITextField) { }
   @objc func newPasswordTextFieldDidReturn(_ sender: UITextField) { }
   @objc func newPasswordConfirmationTextFieldChanged(_ sender: UITextField) { }
